renamed base modal css class to frame, and renamed modalHeader to modal-header
