+++ components/modal/style.scss
@@ -1,32 +1,34 @@
 .components-modal {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  right: auto;
-  bottom: auto;
-  width: auto;
-  max-width: 90%;
-  max-height: 90%;
-  min-height: 50%;
-  border: 0;
-  border-radius: 0;
-  margin-right: -50%;
-  transform: translate(-50%, -50%);
-  background-color: #fff;
-  outline: none;
-  box-shadow: 0 3px 25px #000;
+  &__frame {
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    right: auto;
+    bottom: auto;
+    width: auto;
+    max-width: 90%;
+    max-height: 90%;
+    min-height: 50%;
+    border: 0;
+    border-radius: 0;
+    margin-right: -50%;
+    transform: translate(-50%, -50%);
+    background-color: #fff;
+    outline: none;
+    box-shadow: 0 3px 25px #000;
 
-  // In small screens the content needs to be full width.
-  @media ( max-width: #{ ( $break-small ) } ) {
-    position: fixed;
-    top: 0;
-    bottom: 0;
-    right: 0;
-    left: 0;
-    margin: 0;
-    transform: initial;
-    max-width: none;
-    max-height: none;
+    // In small screens the content needs to be full width.
+    @media (max-width: #{ ( $break-small ) }) {
+      position: fixed;
+      top: 0;
+      bottom: 0;
+      right: 0;
+      left: 0;
+      margin: 0;
+      transform: initial;
+      max-width: none;
+      max-height: none;
+    }
   }
 
   &__screen-overlay {
